requirement.txt.txt
tk
threading
socket
pycrypto
pycryptodome